Moves access modal to component
