+++ apps/judicial-system/web/src/routes/Prosecutor/InvestigationCase/HearingArrangements/HearingArrangements.tsx
@@ -38,7 +38,6 @@ import PageHeader from '@island.is/judicial-system-web/src/components/PageHeader
 import { Box, Input, Text } from '@island.is/island-ui/core'
 import { isHearingArrangementsStepValidIC } from '@island.is/judicial-system-web/src/utils/validate'
 import { formatDateForServer } from '@island.is/judicial-system-web/src/utils/hooks/useCase'
-import type { User } from '@island.is/judicial-system/types'
 import * as constants from '@island.is/judicial-system/consts'
 
 import RequestCourtDate from '../../SharedComponents/RequestCourtDate/RequestCourtDate'
@@ -70,11 +69,6 @@ const HearingArrangements = () => {
     isNotificationModalVisible,
     setIsNotificationModalVisible,
   ] = useState<boolean>(false)
-  const [
-    isProsecutorAccessModalVisible,
-    setIsProsecutorAccessModalVisible,
-  ] = useState<boolean>(false)
-  const [substituteProsecutor, setSubstituteProsecutor] = useState<User>()
 
   const handleNextButtonClick = async () => {
     if (!workingCase) {
@@ -106,21 +100,6 @@ const HearingArrangements = () => {
     }
   }
 
-  const setProsecutor = async (prosecutor: User) => {
-    if (workingCase) {
-      return setAndSendToServer(
-        [
-          {
-            prosecutorId: prosecutor.id,
-            force: true,
-          },
-        ],
-        workingCase,
-        setWorkingCase,
-      )
-    }
-  }
-
   const handleCourtChange = (court: Institution) => {
     if (workingCase) {
       setAndSendToServer(
@@ -138,26 +117,6 @@ const HearingArrangements = () => {
     }
 
     return false
-  }
-
-  const handleProsecutorChange = (prosecutor: User) => {
-    if (!workingCase) {
-      return false
-    }
-
-    const isRemovingCaseAccessFromSelf =
-      user?.id !== workingCase.creatingProsecutor?.id
-
-    if (workingCase.isHeightenedSecurityLevel && isRemovingCaseAccessFromSelf) {
-      setSubstituteProsecutor(prosecutor)
-      setIsProsecutorAccessModalVisible(true)
-
-      return false
-    }
-
-    setProsecutor(prosecutor)
-
-    return true
   }
 
   return (
@@ -187,9 +146,7 @@ const HearingArrangements = () => {
             <Box component="section" marginBottom={7}>
               <CaseInfo workingCase={workingCase} userRole={user.role} />
             </Box>
-            <ProsecutorSectionHeightenedSecurity
-              onChange={handleProsecutorChange}
-            />
+            <ProsecutorSectionHeightenedSecurity />
             <Box component="section" marginBottom={5}>
               <SelectCourt
                 workingCase={workingCase}
@@ -288,27 +245,6 @@ const HearingArrangements = () => {
                   ? formatMessage(errors.sendNotification)
                   : undefined
               }
-            />
-          )}
-          {isProsecutorAccessModalVisible && (
-            <Modal
-              title={formatMessage(m.prosecutorAccessModal.heading)}
-              text={formatMessage(m.prosecutorAccessModal.text)}
-              primaryButtonText={formatMessage(
-                m.prosecutorAccessModal.primaryButtonText,
-              )}
-              secondaryButtonText={formatMessage(
-                m.prosecutorAccessModal.secondaryButtonText,
-              )}
-              handlePrimaryButtonClick={async () => {
-                if (substituteProsecutor) {
-                  await setProsecutor(substituteProsecutor)
-                  router.push(constants.CASES_ROUTE)
-                }
-              }}
-              handleSecondaryButtonClick={() => {
-                setIsProsecutorAccessModalVisible(false)
-              }}
             />
           )}
         </>
