+++ apps/judicial-system/web/src/routes/Prosecutor/SharedComponents/ProsecutorSection/ProsecutorSectionHeightenedSecurity.tsx
@@ -1,35 +1,72 @@
-import React, { useContext } from 'react'
+import React, { useContext, useState } from 'react'
 import { useIntl } from 'react-intl'
+import { useRouter } from 'next/router'
 
 import { Box, Checkbox } from '@island.is/island-ui/core'
 import { User } from '@island.is/judicial-system/types'
+import * as constants from '@island.is/judicial-system/consts'
 import { FormContext } from '@island.is/judicial-system-web/src/components/FormProvider/FormProvider'
 import { UserContext } from '@island.is/judicial-system-web/src/components/UserProvider/UserProvider'
-import { BlueBox } from '@island.is/judicial-system-web/src/components'
+import { BlueBox, Modal } from '@island.is/judicial-system-web/src/components'
 import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import ProsecutorSectionHeading from './ProsecutorSectionHeading'
 import ProsecutorSelection from './ProsecutorSelection'
 import { strings } from './ProsecutorSectionHeightenedSecurity.strings'
 
-interface Props {
-  onChange: (prosecutor: User) => boolean
-}
-
-const ProsecutorSectionHeightenedSecurity: React.FC<Props> = (props) => {
-  const { onChange } = props
-
+const ProsecutorSectionHeightenedSecurity: React.FC = () => {
   const { formatMessage } = useIntl()
+  const router = useRouter()
 
   const { workingCase, setWorkingCase } = useContext(FormContext)
   const { user } = useContext(UserContext)
+  const [substituteProsecutor, setSubstituteProsecutor] = useState<User>()
+  const [
+    isProsecutorAccessModalVisible,
+    setIsProsecutorAccessModalVisible,
+  ] = useState<boolean>(false)
   const { setAndSendToServer } = useCase()
+
+  const setProsecutor = async (prosecutor: User) => {
+    if (workingCase) {
+      return setAndSendToServer(
+        [
+          {
+            prosecutorId: prosecutor.id,
+            force: true,
+          },
+        ],
+        workingCase,
+        setWorkingCase,
+      )
+    }
+  }
+
+  const handleProsecutorChange = (prosecutor: User) => {
+    if (!workingCase) {
+      return false
+    }
+
+    const isRemovingCaseAccessFromSelf =
+      user?.id !== workingCase.creatingProsecutor?.id
+
+    if (workingCase.isHeightenedSecurityLevel && isRemovingCaseAccessFromSelf) {
+      setSubstituteProsecutor(prosecutor)
+      setIsProsecutorAccessModalVisible(true)
+
+      return false
+    }
+
+    setProsecutor(prosecutor)
+
+    return true
+  }
 
   return (
     <Box component="section" marginBottom={5}>
       <ProsecutorSectionHeading />
       <BlueBox>
         <Box marginBottom={2}>
-          <ProsecutorSelection onChange={onChange} />
+          <ProsecutorSelection onChange={handleProsecutorChange} />
         </Box>
         <Checkbox
           name="isHeightenedSecurityLevel"
@@ -56,6 +93,27 @@ const ProsecutorSectionHeightenedSecurity: React.FC<Props> = (props) => {
           filled
         />
       </BlueBox>
+      {isProsecutorAccessModalVisible && (
+        <Modal
+          title={formatMessage(strings.accessModalTitle)}
+          text={formatMessage(strings.accessModalText)}
+          primaryButtonText={formatMessage(
+            strings.accessModalPrimaryButtonText,
+          )}
+          secondaryButtonText={formatMessage(
+            strings.accessModalSecondaryButtonText,
+          )}
+          handlePrimaryButtonClick={async () => {
+            if (substituteProsecutor) {
+              await setProsecutor(substituteProsecutor)
+              router.push(constants.CASES_ROUTE)
+            }
+          }}
+          handleSecondaryButtonClick={() => {
+            setIsProsecutorAccessModalVisible(false)
+          }}
+        />
+      )}
     </Box>
   )
 }
